feat(core): add core and node entry points (#172)
