+++ .changeset/core-node-split.md
@@ -1,0 +1,5 @@
+---
+'@lapidist/design-lint': minor
+---
+
+add core entry point and move node helpers under new node namespace
