+++ src/core/cache-manager.ts
@@ -1,7 +1,8 @@
 import { stat, writeFile } from 'node:fs/promises';
 import type { CacheProvider } from './cache-provider.js';
-import type { LintResult, LintMessage, Fix } from './types.js';
+import type { LintResult } from './types.js';
 import type { LintDocument } from './document-source.js';
+import { applyFixes } from './apply-fixes.js';
 
 export class CacheManager {
   constructor(
@@ -75,24 +76,4 @@ export class CacheManager {
       await this.cache.save();
     }
   }
-}
-
-export function applyFixes(text: string, messages: LintMessage[]): string {
-  const fixes: Fix[] = messages
-    .filter((m): m is LintMessage & { fix: Fix } => m.fix !== undefined)
-    .map((m) => m.fix);
-  if (fixes.length === 0) return text;
-  fixes.sort((a, b) => a.range[0] - b.range[0]);
-  const filtered: Fix[] = [];
-  let lastEnd = -1;
-  for (const f of fixes) {
-    if (f.range[0] < lastEnd) continue;
-    filtered.push(f);
-    lastEnd = f.range[1];
-  }
-  for (let i = filtered.length - 1; i >= 0; i--) {
-    const f = filtered[i];
-    text = text.slice(0, f.range[0]) + f.text + text.slice(f.range[1]);
-  }
-  return text;
 }
