+++ src/core/index.ts
@@ -1,0 +1,26 @@
+export { Linter, type Config } from './linter.js';
+export { applyFixes } from './apply-fixes.js';
+export { Runner } from './runner.js';
+export type { DocumentSource, LintDocument } from './document-source.js';
+export type { PluginLoader, LoadedPlugin } from './plugin-loader.js';
+export type { CacheProvider, CacheEntry } from './cache-provider.js';
+export type {
+  LintResult,
+  LintMessage,
+  RuleModule,
+  RuleContext,
+  RuleListener,
+  DesignTokens,
+  PluginModule,
+  CSSDeclaration,
+  Fix,
+} from './types.js';
+export {
+  matchToken,
+  closestToken,
+  extractVarName,
+  mergeTokens,
+  normalizeTokens,
+  type TokenPattern,
+  type NormalizedTokens,
+} from './token-utils.js';
