+++ src/node/index.ts
@@ -1,0 +1,4 @@
+export { FileSource } from './file-source.js';
+export { createFileDocument } from './file-document.js';
+export { NodePluginLoader } from './plugin-loader.js';
+export { NodeCacheProvider } from './node-cache-provider.js';
