+++ tests/core/linter-integration.test.ts
@@ -3,7 +3,7 @@ import assert from 'node:assert/strict';
 import { promises as fs } from 'node:fs';
 import path from 'node:path';
 import { Linter } from '../../src/core/linter.ts';
-import { FileSource } from '../../src/core/file-source.ts';
+import { FileSource } from '../../src/node/file-source.ts';
 
 void test('Linter integrates registry, parser and trackers', async () => {
   const linter = new Linter(
